This patch makes the TwitterStreamConsumer 's closed flag volatile.
Trying to work out why when the '.close()' method is called the variable
seems to be set correctly (according to the debugger) but the while loop in
run() is not being exited.  If I manually step over the test in the while loop
within the debugger it exits, suggesting an issue with the 'visibility' of the
variable to the TwitterStreamConsumer Thread.